backend/requirements.txt
appnope==0.1.2
APScheduler==3.7.0
asgiref==3.3.4
backcall==0.2.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
dj-rest-auth==2.1.9
Django==3.2.4
django-allauth==0.45.0
django-apscheduler==0.6.0
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mysqlclient==2.0.3
oauthlib==3.1.1
packaging==21.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
PyMySQL==1.0.2
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
traitlets==5.0.5
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.6
wcwidth==0.2.5
